feat: adds popular regions section back to region list
