--- app/components/UI/Ramp/Deposit/components/RegionModal/index.ts
@@ -1,1 +1,0 @@
-export { default } from './RegionModal';
